start adding tests
