+++ xray/utils/sarifutils.go
@@ -2,6 +2,7 @@ package utils
 
 import (
 	"fmt"
+	"github.com/jfrog/gofrog/datastructures"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -295,4 +296,41 @@ func GetInvocationWorkingDirectory(invocation *sarif.Invocation) string {
 		return *invocation.WorkingDirectory.URI
 	}
 	return ""
+}
+
+func getRunWithDummyResults(results ...*sarif.Result) *sarif.Run {
+	run := sarif.NewRunWithInformationURI("", "")
+	ids := datastructures.MakeSet[string]()
+	for _, result := range results {
+		if !ids.Exists(*result.RuleID) {
+			run.Tool.Driver.Rules = append(run.Tool.Driver.Rules, sarif.NewRule(*result.RuleID))
+			ids.Add(*result.RuleID)
+		}
+	}
+	return run.WithResults(results)
+}
+
+func getDummyPassingResult(ruleId string) *sarif.Result {
+	kind := "pass"
+	return &sarif.Result{
+		Kind:   &kind,
+		RuleID: &ruleId,
+	}
+}
+
+func GetDummyResultWithOneLocation(fileName string, startLine, startCol int, snippet, ruleId string, level string) *sarif.Result {
+	return &sarif.Result{
+		Locations: []*sarif.Location{
+			{
+				PhysicalLocation: &sarif.PhysicalLocation{
+					ArtifactLocation: &sarif.ArtifactLocation{URI: &fileName},
+					Region: &sarif.Region{
+						StartLine:   &startLine,
+						StartColumn: &startCol,
+						Snippet:     &sarif.ArtifactContent{Text: &snippet}}},
+			},
+		},
+		Level:  &level,
+		RuleID: &ruleId,
+	}
 }
